[refactor] 메서드의 순서 통일
- 메서드의 순서를 통일해 각 Manager의 역할을 알기 쉽게 함
